pyproject.toml
[tool.poetry]
name = "hermes"
version = "0.1.0"
description = "A simple app to graph wifi speed over time"
authors = ["Sylte <sylteandreas7@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
influxdb-client = "^1.29.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
